refactor: dissolve data-frames data wrapper onto direct @wystack/client
