+++ packages/app/src/app/insights/[insightId]/_components/sections/DataSourcesSection.tsx
@@ -1,7 +1,9 @@
 import { JoinFlowModal } from "@/components/visualizations/JoinFlowModal";
-import { useDataFrames, useInsightMutations } from "@/data";
+import { useInsightMutations } from "@/data";
+import { api } from "@/wystack/api";
 import type { DataTable, Field, Insight } from "@dashframe/types";
 import { JoinTypeIcon } from "@dashframe/ui";
+import { useQuery } from "@wystack/client";
 import {
   ItemList,
   Section,
@@ -31,7 +33,7 @@ export const DataSourcesSection = memo(function DataSourcesSection({
   allTableFields,
 }: DataSourcesSectionProps) {
   const [isJoinFlowOpen, setIsJoinFlowOpen] = useState(false);
-  const { data: allDataFrameEntries = [] } = useDataFrames();
+  const { data: allDataFrameEntries = [] } = useQuery(api.listDataFrames);
   const { update: updateInsight } = useInsightMutations();
 
   // Remove join handler
