+++ packages/app/src/components/preview-diff/usePreviewComputeFill.test.ts
@@ -45,7 +45,7 @@ const { mockGetDataFrame, mockGetDataTable } = vi.hoisted(() => ({
   mockGetDataTable: vi.fn(),
 }));
 
-vi.mock("@/data", () => ({
+vi.mock("@/lib/data-access/data-frames", () => ({
   getDataFrame: mockGetDataFrame,
 }));
