+++ packages/app/src/components/preview-diff/usePreviewComputeFill.ts
@@ -27,7 +27,7 @@
  */
 
 import { useDuckDB } from "@/components/providers/DuckDBProvider";
-import { getDataFrame } from "@/data";
+import { getDataFrame } from "@/lib/data-access/data-frames";
 import { getDataTable } from "@/lib/data-access/data-tables";
 import { buildInsightSQL, ensureTableLoaded } from "@dashframe/engine-browser";
 import type {
