+++ packages/app/src/hooks/useInsightPagination.test.ts
@@ -51,7 +51,7 @@ vi.mock("@/components/providers/DuckDBProvider", () => ({
   useDuckDB: () => mockUseDuckDB(),
 }));
 
-vi.mock("@/data", () => ({
+vi.mock("@/lib/data-access/data-frames", () => ({
   getDataFrame: mockGetDataFrame,
 }));
