+++ packages/app/src/hooks/useInsightPagination.test.tsx
@@ -38,7 +38,7 @@ const { mockGetDataFrame, mockGetDataTable } = vi.hoisted(() => ({
   mockGetDataTable: vi.fn(),
 }));
 
-vi.mock("@/data", () => ({
+vi.mock("@/lib/data-access/data-frames", () => ({
   getDataFrame: mockGetDataFrame,
 }));
