+++ packages/app/src/hooks/useInsightPagination.ts
@@ -1,5 +1,5 @@
 import { useDuckDB } from "@/components/providers/DuckDBProvider";
-import { getDataFrame } from "@/data";
+import { getDataFrame } from "@/lib/data-access/data-frames";
 import { getDataTable } from "@/lib/data-access/data-tables";
 import type { EffectiveParams } from "@dashframe/engine";
 import {
