+++ packages/app/src/hooks/useInsightView.test.tsx
@@ -42,7 +42,7 @@ const { mockGetDataFrame, mockGetDataTable } = vi.hoisted(() => ({
   mockGetDataTable: vi.fn(),
 }));
 
-vi.mock("@/data", () => ({
+vi.mock("@/lib/data-access/data-frames", () => ({
   getDataFrame: mockGetDataFrame,
 }));
