+++ packages/app/src/hooks/useInsightView.ts
@@ -1,6 +1,6 @@
 import { useChartEngine } from "@/components/providers/ChartEngineProvider";
 import { useDuckDB } from "@/components/providers/DuckDBProvider";
-import { getDataFrame } from "@/data";
+import { getDataFrame } from "@/lib/data-access/data-frames";
 import { getDataTable } from "@/lib/data-access/data-tables";
 import type { EffectiveParams } from "@dashframe/engine";
 import {
