+++ packages/app/src/lib/remote-table-materialization.ts
@@ -1,4 +1,8 @@
-import { addDataFrameEntry, getDataFrameEntry, removeDataFrame } from "@/data";
+import {
+  addDataFrameEntry,
+  getDataFrameEntry,
+  removeDataFrame,
+} from "@/lib/data-access/data-frames";
 import { getDataTable, updateDataTable } from "@/lib/data-access/data-tables";
 import { DataFrame, deleteArrowData } from "@dashframe/engine-browser";
 import { getFieldSensitivity, type Field, type UUID } from "@dashframe/types";
